Rename src/contrib to src/minizip
Directory name "contrib" is too generic and doesn't really fit,
so rename it to "minizip", which is much clearer.

Update all references to the directory.

Remove the directory from the include path and add it to the #include
statement.
