+++ src/CMakeLists.txt
@@ -1,7 +1,6 @@
-file(GLOB SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/*.c ${CMAKE_CURRENT_SOURCE_DIR}/contrib/*.c)
+file(GLOB SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/*.c ${CMAKE_CURRENT_SOURCE_DIR}/minizip/*.c)
 add_executable(trrntzip ${SOURCES})
 target_compile_definitions(trrntzip PRIVATE "TZ_VERSION=\"${PROJECT_VERSION}\"")
-target_include_directories(trrntzip BEFORE PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/contrib)
 target_link_libraries(trrntzip m ZLIB::ZLIB)
-set_property(SOURCE ${CMAKE_CURRENT_SOURCE_DIR}/contrib/unzip.c APPEND PROPERTY COMPILE_DEFINITIONS UNZ_MAXFILENAMEINZIP=1024)
+set_property(SOURCE ${CMAKE_CURRENT_SOURCE_DIR}/minizip/unzip.c APPEND PROPERTY COMPILE_DEFINITIONS UNZ_MAXFILENAMEINZIP=1024)
 install(TARGETS trrntzip EXPORT ${PROJECT_NAME}-targets DESTINATION bin)
